Move PipSession to network.session
