+++ src/pip/_internal/legacy_resolve.py
@@ -44,7 +44,7 @@ if MYPY_CHECK_RUNNING:
     from pip._vendor import pkg_resources
 
     from pip._internal.distributions import AbstractDistribution
-    from pip._internal.download import PipSession
+    from pip._internal.network.session import PipSession
     from pip._internal.index import PackageFinder
     from pip._internal.operations.prepare import RequirementPreparer
     from pip._internal.req.req_install import InstallRequirement
