+++ src/pip/_internal/operations/prepare.py
@@ -32,8 +32,8 @@ if MYPY_CHECK_RUNNING:
     from typing import Optional
 
     from pip._internal.distributions import AbstractDistribution
-    from pip._internal.download import PipSession
     from pip._internal.index import PackageFinder
+    from pip._internal.network.session import PipSession
     from pip._internal.req.req_install import InstallRequirement
     from pip._internal.req.req_tracker import RequirementTracker
