+++ tests/unit/test_index.py
@@ -4,7 +4,6 @@ import pytest
 from pip._vendor.packaging.specifiers import SpecifierSet
 
 from pip._internal.collector import LinkCollector
-from pip._internal.download import PipSession
 from pip._internal.index import (
     CandidateEvaluator,
     CandidatePreferences,
@@ -21,6 +20,7 @@ from pip._internal.models.link import Link
 from pip._internal.models.search_scope import SearchScope
 from pip._internal.models.selection_prefs import SelectionPreferences
 from pip._internal.models.target_python import TargetPython
+from pip._internal.network.session import PipSession
 from pip._internal.pep425tags import get_supported
 from pip._internal.utils.hashes import Hashes
 from tests.lib import CURRENT_PY_VERSION_INFO
