+++ tests/unit/test_network_session.py
@@ -1,8 +1,9 @@
+import logging
+
 import pytest
 
 import pip
-from pip._internal.download import PipSession
-from pip._internal.network.requests import CI_ENVIRONMENT_VARIABLES
+from pip._internal.network.session import CI_ENVIRONMENT_VARIABLES, PipSession
 
 
 def get_user_agent():
@@ -46,3 +47,173 @@ def test_user_agent__ci(monkeypatch, name, expected_like_ci):
 def test_user_agent_user_data(monkeypatch):
     monkeypatch.setenv("PIP_USER_AGENT_USER_DATA", "some_string")
     assert "some_string" in PipSession().headers["User-Agent"]
+
+
+class TestPipSession:
+
+    def test_cache_defaults_off(self):
+        session = PipSession()
+
+        assert not hasattr(session.adapters["http://"], "cache")
+        assert not hasattr(session.adapters["https://"], "cache")
+
+    def test_cache_is_enabled(self, tmpdir):
+        session = PipSession(cache=tmpdir.joinpath("test-cache"))
+
+        assert hasattr(session.adapters["https://"], "cache")
+
+        assert (session.adapters["https://"].cache.directory ==
+                tmpdir.joinpath("test-cache"))
+
+    def test_http_cache_is_not_enabled(self, tmpdir):
+        session = PipSession(cache=tmpdir.joinpath("test-cache"))
+
+        assert not hasattr(session.adapters["http://"], "cache")
+
+    def test_insecure_host_adapter(self, tmpdir):
+        session = PipSession(
+            cache=tmpdir.joinpath("test-cache"),
+            trusted_hosts=["example.com"],
+        )
+
+        assert "https://example.com/" in session.adapters
+        # Check that the "port wildcard" is present.
+        assert "https://example.com:" in session.adapters
+        # Check that the cache isn't enabled.
+        assert not hasattr(session.adapters["https://example.com/"], "cache")
+
+    def test_add_trusted_host(self):
+        # Leave a gap to test how the ordering is affected.
+        trusted_hosts = ['host1', 'host3']
+        session = PipSession(trusted_hosts=trusted_hosts)
+        insecure_adapter = session._insecure_adapter
+        prefix2 = 'https://host2/'
+        prefix3 = 'https://host3/'
+        prefix3_wildcard = 'https://host3:'
+
+        # Confirm some initial conditions as a baseline.
+        assert session.pip_trusted_origins == [
+            ('host1', None), ('host3', None)
+        ]
+        assert session.adapters[prefix3] is insecure_adapter
+        assert session.adapters[prefix3_wildcard] is insecure_adapter
+
+        assert prefix2 not in session.adapters
+
+        # Test adding a new host.
+        session.add_trusted_host('host2')
+        assert session.pip_trusted_origins == [
+            ('host1', None), ('host3', None), ('host2', None)
+        ]
+        # Check that prefix3 is still present.
+        assert session.adapters[prefix3] is insecure_adapter
+        assert session.adapters[prefix2] is insecure_adapter
+
+        # Test that adding the same host doesn't create a duplicate.
+        session.add_trusted_host('host3')
+        assert session.pip_trusted_origins == [
+            ('host1', None), ('host3', None), ('host2', None)
+        ], 'actual: {}'.format(session.pip_trusted_origins)
+
+        session.add_trusted_host('host4:8080')
+        prefix4 = 'https://host4:8080/'
+        assert session.pip_trusted_origins == [
+            ('host1', None), ('host3', None),
+            ('host2', None), ('host4', 8080)
+        ]
+        assert session.adapters[prefix4] is insecure_adapter
+
+    def test_add_trusted_host__logging(self, caplog):
+        """
+        Test logging when add_trusted_host() is called.
+        """
+        trusted_hosts = ['host0', 'host1']
+        session = PipSession(trusted_hosts=trusted_hosts)
+        with caplog.at_level(logging.INFO):
+            # Test adding an existing host.
+            session.add_trusted_host('host1', source='somewhere')
+            session.add_trusted_host('host2')
+            # Test calling add_trusted_host() on the same host twice.
+            session.add_trusted_host('host2')
+
+        actual = [(r.levelname, r.message) for r in caplog.records]
+        # Observe that "host0" isn't included in the logs.
+        expected = [
+            ('INFO', "adding trusted host: 'host1' (from somewhere)"),
+            ('INFO', "adding trusted host: 'host2'"),
+            ('INFO', "adding trusted host: 'host2'"),
+        ]
+        assert actual == expected
+
+    def test_iter_secure_origins(self):
+        trusted_hosts = ['host1', 'host2', 'host3:8080']
+        session = PipSession(trusted_hosts=trusted_hosts)
+
+        actual = list(session.iter_secure_origins())
+        assert len(actual) == 9
+        # Spot-check that SECURE_ORIGINS is included.
+        assert actual[0] == ('https', '*', '*')
+        assert actual[-3:] == [
+            ('*', 'host1', '*'),
+            ('*', 'host2', '*'),
+            ('*', 'host3', 8080)
+        ]
+
+    def test_iter_secure_origins__trusted_hosts_empty(self):
+        """
+        Test iter_secure_origins() after passing trusted_hosts=[].
+        """
+        session = PipSession(trusted_hosts=[])
+
+        actual = list(session.iter_secure_origins())
+        assert len(actual) == 6
+        # Spot-check that SECURE_ORIGINS is included.
+        assert actual[0] == ('https', '*', '*')
+
+    @pytest.mark.parametrize(
+        'location, trusted, expected',
+        [
+            ("http://pypi.org/something", [], False),
+            ("https://pypi.org/something", [], True),
+            ("git+http://pypi.org/something", [], False),
+            ("git+https://pypi.org/something", [], True),
+            ("git+ssh://git@pypi.org/something", [], True),
+            ("http://localhost", [], True),
+            ("http://127.0.0.1", [], True),
+            ("http://example.com/something/", [], False),
+            ("http://example.com/something/", ["example.com"], True),
+            # Try changing the case.
+            ("http://eXample.com/something/", ["example.cOm"], True),
+            # Test hosts with port.
+            ("http://example.com:8080/something/", ["example.com"], True),
+            # Test a trusted_host with a port.
+            ("http://example.com:8080/something/", ["example.com:8080"], True),
+            ("http://example.com/something/", ["example.com:8080"], False),
+            (
+                "http://example.com:8888/something/",
+                ["example.com:8080"],
+                False
+            ),
+        ],
+    )
+    def test_is_secure_origin(self, caplog, location, trusted, expected):
+        class MockLogger(object):
+            def __init__(self):
+                self.called = False
+
+            def warning(self, *args, **kwargs):
+                self.called = True
+
+        session = PipSession(trusted_hosts=trusted)
+        actual = session.is_secure_origin(location)
+        assert actual == expected
+
+        log_records = [(r.levelname, r.message) for r in caplog.records]
+        if expected:
+            assert not log_records
+            return
+
+        assert len(log_records) == 1
+        actual_level, actual_message = log_records[0]
+        assert actual_level == 'WARNING'
+        assert 'is not a trusted or secure host' in actual_message
