+++ tests/unit/test_req.py
@@ -11,7 +11,6 @@ from pip._vendor.packaging.markers import Marker
 from pip._vendor.packaging.requirements import Requirement
 
 from pip._internal.commands import create_command
-from pip._internal.download import PipSession
 from pip._internal.exceptions import (
     HashErrors,
     InstallationError,
@@ -19,6 +18,7 @@ from pip._internal.exceptions import (
     PreviousBuildDirError,
 )
 from pip._internal.legacy_resolve import Resolver
+from pip._internal.network.session import PipSession
 from pip._internal.operations.prepare import RequirementPreparer
 from pip._internal.req import InstallRequirement, RequirementSet
 from pip._internal.req.constructors import (
